Minor fixes

+++ app/crawiot_config.h
@@ -6,7 +6,7 @@ struct Config {
     const char* HostName;
     const char* DeviceIp;
     const int FirstSpeedDegree;
-    const int WheelRadius;
+    const float WheelRadius;
 };
 
 #endif //APP_CRAWIOT_CONFIG_H
